First commit of Casino Game.
This ASP.NET application is a Casino Game based on exercise 34 in Bob Tabor's ASP.NET / C# course on DevU.com. The images were provided by DevU.com but the solution is mine. The exercise did not require the code to be built using an OO design, but I decided to nonetheless.

The game gives the user $100 to start and allows them to bet money, pull a lever and win or lose money based on the rules on the page.